requirements.txt
yt-dlp>=2024.10.7
anthropic>=0.40.0
PyYAML>=6.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
Pillow>=10.0.0
